Add GhostColorForm for Rosa File Customization
Implemented GhostColorForm to allow users to choose a name and color for their Rosa file. The form enables the selection of a color from predefined options and input for custom file naming. Enhanced the encryption process to embed the chosen color as metadata within the encrypted file. This update paves the way for future decryption features that will utilize the color metadata for custom animations and user experience enhancements. Also changed some details on the Main Menu UI.